Refactored HandlerService objects into their own file
